DishHub/requirements.txt
apturl==0.5.2
asgiref==3.8.1
bcrypt==3.2.0
beniget==0.4.1
blessed==1.20.0
blinker==1.4
Brlapi==0.8.3
catalogue==2.0.10
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contexttimer==0.3.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distlib==0.3.8
distro==1.7.0
distro-info===1.1build1
Django==3.2.4
django-jazzmin==3.0.0
djangorestframework==3.12.4
duplicity==0.8.21
fasteners==0.14.1
filelock==3.15.4
future==0.18.2
gast==0.5.2
gitdb==4.0.11
GitPython==3.1.42
httplib2==0.20.2
humanfriendly==10.0
idna==3.3
ifaddr==0.1.7
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.4
numpy-stl==2.8.0
nvm==1.0.21
oauthlib==3.2.0
olefile==0.46
PAM==0.4.2
pandas==2.2.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.2
ply==3.11
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyserial==3.5
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-utils==2.2.0
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
scipy==1.8.0
SecretStorage==3.3.1
sentry-sdk==1.4.3
Shapely==1.8.0
six==1.16.0
smmap==5.0.1
sqlparse==0.4.2
srsly==2.4.8
systemd-python==234
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.26.3
wadllib==1.3.6
wcwidth==0.2.13
xdg==5
xkit==0.0.0
zeroconf==0.38.3
zipp==1.0.0
